リファクタリング 主に、スキル説明のパラメータ変換について
