+++ components/skill-window/sortable-skill-card.tsx
@@ -1,0 +1,37 @@
+"use client"
+
+import type React from "react"
+import { useSortable } from "@dnd-kit/sortable"
+import { CSS } from "@dnd-kit/utilities"
+
+interface SortableSkillCardProps {
+  id: string
+  children: React.ReactNode
+}
+
+export function SortableSkillCard({ id, children }: SortableSkillCardProps) {
+  const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
+    id,
+  })
+
+  const style = {
+    transform: CSS.Transform.toString(transform),
+    transition,
+    opacity: isDragging ? 0.5 : 1,
+    zIndex: isDragging ? 100 : 1,
+    position: "relative" as const,
+  }
+
+  return (
+    <div
+      ref={setNodeRef}
+      style={style}
+      {...attributes}
+      {...listeners}
+      className={`touch-manipulation ${isDragging ? "dragging-card" : ""}`}
+      onContextMenu={(e) => e.preventDefault()}
+    >
+      {children}
+    </div>
+  )
+}
